Create the scanner
This commit creates a scanner.  A scanner knows how to take a source and
read it with the transformations requested by the options, returning each
character (as a parser.AugChar) one at a time.  The scanner is also
capable of detecting and utilizing the line ending style of the file, by
picking up the first newline, carriage return, or carriage return/newline
pair.  Finally, the scanner implements arbitrary character pushback, which
will simplify backtracking in the lexer.

This code also brings in some additional options, namely, the TabStop
option (defaults to 8) for setting the tab stop, and an Options.Advance()
method that advances a location based on the character.  This could allow
more advanced column calculations in the future, taking into account the
Unicode combining characters, for instance.