Create test utilities
This commit creates the test utilities AssertPtrEqual and
AssertPtrNotEqual.  There are limititations of this type of approach--in
particular, optimizations can cause two identical objects to be optimized
out--but this is handy for, e.g., function pointer comparisons.

This commit uses these utilities in the lineendings test suite for the
scanner.