Implement recognizer for numbers
This commit includes the recognizer for numbers.  This recognizer is
capable of extracting both integers and floats, handling '.', 'e', and
an optional sign ('+' or '-') after 'e'.  Binary, octal, and hex bases
can be used for integers.  The values are math/big.Int and math/big.Float,
to be similar to the capabilities of Python in this space.