Add Keywords and Norm
This commit includes the simple object Keywords, with Copy(), Add(), and
Remove() methods.  This will be used to differentiate keywords from
regular identifiers by the lexer.  This commit also adds a Norm field to
the profile, which will contain the normalization to use on identifiers;
by parameterizing the normalization, this allows us to, in future,
decouple Hydra's identifiers from golang's native unicode database.