Recognize identifiers
This commit contains the recognizer for identifiers.  This recognizer is a
little more complicated, because it has to take into account the
possibility of a Python-style string flag (e.g., 'b"foo"' for a byte
string); it accomplishes this by initializing a string recognizer, and
throwing it away when a character comes along that is not a valid string
flag.  The identifier recognizer also has to apply Unicode normalization
rules to the recognized identifier.

The identifier recognizer also recognizes keywords by looking them up in
the profile's Keywords map.