Recognize comments
This commit includes a recognizer for comments.  For regular comments,
everything up to the newline (or EOF) is simply consumed; however, doc
comments--which are comments with the comment character doubled--will have
leading whitespace stripped, then returned as a <DocComment> token.

This commit also sneaks in a minor change to symbol names to enclose them
in angle brackets, to differentiate them from identifier symbols and,
eventually, operator symbols.