Ensure we don't stop when newlines follow each other
In testing the lexer, it was discovered that if a bare newline appeared
at the beginning of the file being processed, or if a newline followed
another newline, the Next() method would return nil.  This was traced,
finally, to a "break" after generating a newline token; problem is,
newline tokens may be elided in those circumstances.  The correct thing to
do is to continue: if the newline token was not elided, the loop will exit
because the token queue now has a token on it, but if it was elided, the
loop will re-execute and recognize the next token.

This commit includes a failing test for this condition, to prevent future
regressions.