Annotated children
Sometimes the children returned by the Children() call have special
meaning, such as the left-hand-side of a binary operator.  To enable this
information to be called out in a visualization, this commit adds the
Annotated structure, which wraps another Visitable.  The Children() call
is proxied to the wrapped Visitable, while the String() call has a
specified prefix added to it to denote the special relationship.