Segregate parser and create top level
This commit completes the reorganization of the source tree (for now) by
moving the parser code into a segregated subdirectory (the "parser"
subdirectory of the "parser" tree).  It also moves the Filename, Encoding,
and TabStop options into the top-level of the "parser" tree, in
preparation for the future addition of parsing entrypoints that will in
turn utilize the parser implementation in the "parser" subdirectory.